Altered packages. Deleted service package and moved json package directly to network package. Added Javadoc to ActionValues.java.
