+++ app/src/main/java/com/example/se2_projekt_app/networking/json/GenerateJSONObject.java
@@ -1,4 +1,4 @@
-package com.example.se2_projekt_app.networking.services.JSON;
+package com.example.se2_projekt_app.networking.json;
 
 import org.json.JSONObject;
