--- app/src/main/java/com/example/se2_projekt_app/networking/services/JSON/ActionValues.java
@@ -1,15 +1,0 @@
-package com.example.se2_projekt_app.networking.services.JSON;
-
-public enum ActionValues {
-    REGISTERUSER("registerUser"),
-    JOINLOBBY("joinLobby");
-
-    private final String value;
-    ActionValues(String action){
-        this.value = action;
-    }
-
-    public String getValue() {
-        return this.value;
-    }
-}
